refactor: externalize cart button
